check main file exists before delete

+++ scripts/pre_debug_script.py
@@ -4,6 +4,7 @@ from distutils.dir_util import copy_tree
 import shutil, os
 
 copy_tree("src" , "test")
-os.remove("test/main.cpp")
+if os.path.exists("test/main.cpp"):
+    os.remove("test/main.cpp")
 
 print('### Finish Pre-debug script ########################')
